+++ gcs/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Next
 1. Upgrade to use cloud-storage libraries-bom
+1. Add AUTO_RANDOM as new fadvise mode.
 
 ## 3.1.1 - 2025-04-07
 1. PR #1323: Fix a bug where gRPC path was failing due to some of the dependencies not getting shaded.